+++ src/main/java/com/ouaskanas/educonnect/Service/Manager/FrienshipManager.java
@@ -5,9 +5,9 @@ import com.ouaskanas.educonnect.Dao.Entities.Friendship;
 import java.util.List;
 
 public interface FrienshipManager {
-    public List<Friendship> friendshipListforUser(long user_id);
-    public Friendship sendFriendShipRequest(long user_id, long friend_id);
-    public Friendship acceptFriendshipRequest(long friendship_id);
-    public void declineFriendshipRequest(long friendship_id);
+    public List<Friendship> friendshipListforUser(int user_id);
+    public Friendship sendFriendShipRequest(int user_id, int friend_id);
+    public Friendship acceptFriendshipRequest(int friendship_id);
+    public void declineFriendshipRequest(int friendship_id);
 
 }
