+++ src/main/java/com/ouaskanas/educonnect/Service/Manager/PostManager.java
@@ -6,10 +6,10 @@ import com.ouaskanas.educonnect.Dto.PostDto;
 import java.util.List;
 
 public interface PostManager {
-    public Post getPost(long id);
+    public Post getPost(int id);
     public List<Post> getAllPosts();
-    public void deletePost(long id);
+    public void deletePost(int id);
     public Post addPost(PostDto post);
-    public Post updatePost(long id,PostDto post);
+    public Post updatePost(int id,PostDto post);
     public Post getPostByTitle(String title);
 }
